Merchant Ship implemented the 5 tile long ship when placed now produces 1 gold per second
