--- src/WaterWizard.Client/gamescreen/ships/ShipType.cs
@@ -1,20 +1,0 @@
-// ===============================================
-// Autoren-Statistik (automatisch generiert):
-// - maxk2807: 7 Zeilen
-// - Paul: 4 Zeilen
-// 
-// Methoden/Funktionen in dieser Datei (Hauptautor):
-// (Keine Methoden/Funktionen gefunden)
-// ===============================================
-
-namespace WaterWizard.Client.gamescreen.ships;
-
-public enum ShipType
-{
-    DEFAULT,
-    MERCHANT,
-    ONEFIELDSHIP,
-    TWOFIELDSHIP,
-    FOURFIELDSHIP,
-    FIVEFIELDSHIP
-}
