+++ src/WaterWizard.Shared/ShipType.cs
@@ -1,0 +1,11 @@
+namespace WaterWizard.Shared.ShipType;
+
+public enum ShipType
+{
+    DEFAULT,
+    Merchant,
+    ONEFIELDSHIP,
+    TWOFIELDSHIP,
+    FOURFIELDSHIP,
+    FIVEFIELDSHIP
+}
